work in progress

+++ DawgBuilder/Form1.cs
@@ -1,4 +1,4 @@
-﻿using DawgResolver;
+﻿using Dawg;
 
 using System;
 using System.IO;
@@ -10,7 +10,7 @@ namespace Test_DAWG
     public partial class Form1 : Form
     {
 
-        Dictionnaire dico;
+        Dictionnaire dico = new Dictionnaire();
 
         public Form1()
         {
@@ -19,21 +19,17 @@ namespace Test_DAWG
 
         private void Form1_Load(object sender, EventArgs e)
         {
-            dico = new Dictionnaire();
             dico.EtapeAtteinte += dico_EtapeAtteinte;
             dico.Progression += dico_Progression;
-            dico.ChargerDictionnaireAscii(Dictionnaire.NomDicoReel);
-            grpLectureDAWG.Enabled = File.Exists(Dictionnaire.NomDicoDawgODS7);
-            grpAjouterMot.Enabled = grpLectureDAWG.Enabled;
         }
 
         void dico_EtapeAtteinte(string Etape, long TempsExecution)
         {
-            string texte = Etape + TempsExecution/1000.0 + " secondes.";
+            string texte = Etape + TempsExecution / 1000.0 + " secondes.";
 
             ListBox lst;
 
-            switch(dico.TravailEnCours)
+            switch (dico.TravailEnCours)
             {
                 case TravailEnCours.ChargementFichierASCII:
                     lst = lstChargementASCII;
@@ -59,14 +55,6 @@ namespace Test_DAWG
             lst.SelectedIndex = lst.Items.Count - 1;
         }
 
-        private void button1_Click(object sender, EventArgs e)
-        {
-            Cursor.Current = Cursors.WaitCursor;
-            dico.ConstruireDawgEnDeuxTemps();
-            Cursor.Current = Cursors.Default;
-
-        }
-
         void dico_Progression(int Pourcent)
         {
             switch (dico.TravailEnCours)
@@ -76,7 +64,6 @@ namespace Test_DAWG
 
                 case TravailEnCours.CreationDawgEn2Etapes:
                     pgb2etapes.Value = Pourcent;
-                    grpLectureDAWG.Enabled = (Pourcent == 100);
                     grpAjouterMot.Enabled = grpLectureDAWG.Enabled;
                     break;
 
@@ -91,9 +78,16 @@ namespace Test_DAWG
             }
         }
 
-        private void butLectureDAWG_Click(object sender, EventArgs e)
+        private void btnLectureDAWG_Click(object sender, EventArgs e)
         {
-            dico.ChargerFichierDAWG();
+            using (var ofd = new OpenFileDialog() { CheckFileExists = true, RestoreDirectory = true })
+            {
+                if (ofd.ShowDialog() == DialogResult.OK)
+                {
+                  
+                    dico.ChargerFichierDAWG(ofd.FileName);
+                }
+            }
         }
 
         private void butAjouterUnMot_Click(object sender, EventArgs e)
@@ -103,8 +97,36 @@ namespace Test_DAWG
 
         private void button2_Click(object sender, EventArgs e)
         {
-           var words = File.ReadAllLines(Dictionnaire.NomDicoReel).OrderBy(m => m).ToList();
+            if (!File.Exists(Dictionnaire.NomDicoReel)) return;
+            var words = File.ReadAllLines(Dictionnaire.NomDicoReel).OrderBy(m => m).ToList();
             File.WriteAllLines(Dictionnaire.NomDicoReel, words);
+        }
+
+        private void btnReadDicoText_Click(object sender, EventArgs e)
+        {
+            bool dicoLoaded = false;
+            using (var ofd = new OpenFileDialog() { CheckFileExists = true, RestoreDirectory = true })
+            {
+                if (ofd.ShowDialog() == DialogResult.OK)
+                {
+                    Dictionnaire.NomDicoReel = ofd.FileName;
+                    
+
+                    grpLectureDAWG.Enabled = true;
+                    grpAjouterMot.Enabled = grpLectureDAWG.Enabled;
+                    dicoLoaded = dico.ChargerDictionnaireAscii(ofd.FileName);
+                }
+            }
+            btnCreationDawg.Enabled = dicoLoaded;
+        }
+
+        private void btnCreationDawg_Click(object sender, EventArgs e)
+        {
+            Cursor.Current = Cursors.WaitCursor;
+            dico.ConstruireDawgEnDeuxTemps();
+
+            Cursor.Current = Cursors.Default;
+
         }
     }
 }
